nothing of note

+++ src/test.cpp
@@ -18,6 +18,7 @@ inline constexpr T min(const T& input){
     return (T)std::numeric_limits<T>::min();
 }
 
+
 /*
 if the last bit is a 1 (negative, substract) then use & operator
 if the last bit is 0
@@ -26,7 +27,7 @@ if the last bit is 0
 int main(void){
     long hi = -2;
 
-    std::cout << BITS(max(hi)) << N <<  (int)min(hi) << N ; 
+    std::cout << BITS(max(hi)) << N <<  min(hi) << N ; 
     test();
